Core/DBLayer: Few more type corrections
Closes #5871

+++ src/server/game/Server/WorldSocket.cpp
@@ -909,7 +909,7 @@ int WorldSocket::HandleAuthSession (WorldPacket& recvPacket)
     else
     {
         fields = result->Fetch();
-        security = fields[0].GetInt32();
+        security = fields[0].GetUInt8();
     }
 
     // Re-check account ban (same check as in realmd)
